### 2019-11-08 #1875 @whedon
[PRE REVIEW]: ampscan: A lightweight Python package for shape analysis of prosthetics and orthotics
**Submitting author:** @JoshuaSteer (<a href="http://orcid.org/0000-0002-6288-1347">Joshua Steer</a>)
**Repository:** <a href="https://github.com/abel-research/ampscan" target ="_blank">https://github.com/abel-research/ampscan</a>
**Version:** v0.3.0
