### 2019-11-08 #1875 @kyleniemeyer
@openjournals/dev any ideas why the paper isn't compiling? I can't tell from the error message.
