### 2019-11-11 #1875 @JoshuaSteer
@Kevin-Mattheus-Moerman Thanks for picking this up, and great suggestion of Dana (also who I would have suggested). I'll have a look around for potential reviewers, possibly the team behind the polygon mesh processing library (https://github.com/pmp-library/pmp-library)?
