### 2019-11-14 #1875 @alexd103
Thanks for handling this for us @Kevin-Mattheus-Moerman!
Tom Saey from MOBILAB at Thomas More in Belgium might be interested and has domain-relevant experience - again if we can find him on here. https://www.linkedin.com/in/tom-saey-61167232/?originalSubdomain=be
