### 2019-12-05 #1875 @Kevin-Mattheus-Moerman
@gkontoudis @cmavrogiannis @liarokapis would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org) on prosthetic socket design?
