### 2020-01-11 #1875 @Kevin-Mattheus-Moerman
@gkontoudis @cmavrogiannis @liarokapis would you be interested in reviewing this submission for JOSS on prosthetic socket design? christmas_tree
