### 2020-02-01 #2060 @Kevin-Mattheus-Moerman
@JoshuaSteer from the above report it looks like Omar Animashaun contributed but it is not listed as an author or acknowledged in the paper. Can you comment on this? If they are not included as author (e.g. if their contributions were minor enough), perhaps you can still acknowledge other contributors like these in the paper?
