### 2020-02-04 dfafd26
Removed ssm from repo as not directly within scope of ampscan project
