### 2020-02-07 #2060 @Kevin-Mattheus-Moerman
@JoshuaSteer
See my earlier comments and these additional comments.
- [ ] Can you work on adding clear contributing guidelines? (one of mine for example is here: https://github.com/gibbonCode/GIBBON/blob/master/CONTRIBUTING.md you can link to this in the readme). But you can be as brief/extensive as you want. Perhaps you can link to Python style standards. You can search around a bit for what is commonly used. If you intend to have this community grow I would consider adding a [code of conduct](https://opensource.guide/code-of-conduct/) as well.
- [ ] Can you add a simple getting started section? I ran the 3 examples as Jupyter notebooks. This is convenient and works but it does not say this anywhere (unless I missed it). It seems the documentation site just renders these jupyter notebooks. So a section on installation (it is in the readme but the documentation lacks it and could repeat it)  and also a getting started section would be good.
- [ ] Can you add automated tests? E.g. a more advanced (non-Jupyter notebook) implementation. If you do this then adding testing instructions would be good.
- [ ] Check for the typos `vertrx`, and `convienience` in the fundamentals demo.
- [ ] It looks like you migrated to GitHub from GitLab. The documentation still refers to GitLab though so perhaps fix that.
- [ ] Recommendation: Volume computation does not require slicing. And by using slicing it becomes less accurate than it can be (as shown in the coarse/fine comparison). Have you thought about exact volume computation based on Gauss divergence theorem (here is a MATLAB implementation https://github.com/gibbonCode/GIBBON/blob/master/lib/triSurfVolume.m, note for quads you can first convert to triangular faces by slashing the quads). You probably need to close over the top of the scan though (although that should not be too hard).
- [ ] Recommendation: I did not check your smoothing implementation but saw statements on shape changes induced by it. Have you seen this simple smooth method called [HC smoothing](http://www.joerg-vollmer.de/downloads/Improved_Laplacian_Smoothing_of_Noisy_Surface_Meshes.pdf) it helps to reduce "shrinkage" effects.
