### 2020-02-09 e2f2741
Upadated requirements file as pyqt 5.14 fails to install
