+++ requirements.txt
@@ -3,6 +3,6 @@ matplotlib>=2
 scipy==1.2.1
 sphinxcontrib-napoleon==0.7
 vtk>=8
-PyQt5>=5
+PyQt5>=5, <=5.13
 PyPDF2>=1.20
 reportlab>=3
