### 2020-02-09 343fd43
Updated typos in the jupyter notebooks
