+++ docs/exampleDemos.rst
@@ -3,7 +3,7 @@ Tutorials
 
 This section provides a series of workflows that demonstrate the 
 core functionality of ampscan. Some require sample stl 
-files - these can be downloaded from the Gitlab repository `here`_. 
+files - these can be downloaded from the Github repository `here`_. 
 If you are using a git clone you can make the module importable 
 via ``import ampscan`` by navigating to the root directory and running 
 ``pip install -e .``.
