### 2020-02-09 #2060 @JoshuaSteer
@Kevin-Mattheus-Moerman Thanks for all the comments, all great suggestions to improve ampscan.

Quick Q on testing, we've got our whole package linked up to travis CI (https://travis-ci.org/abel-research/ampscan), so I guess you are referring to enabling for someone to run these tests locally on install as well?

Thanks!
