### 2020-02-09 ef1f512
Fixed references in paper and added DOIs
