+++ joss/paper.md
@@ -52,7 +52,7 @@ The typical ampscan process is demonstrated above (Figure 1) and the core functi
 
 - **Graphical User Interface**: This enables visualisation to multiple AmpObjects within a single window, giving access to the automated and manual alignment tools as well as registration. This facilitates the core analysis of the scan data for users who are not experienced Python users. The python version of the GUI is available in the [gui](https://github.com/abel-research/ampscan/tree/master/gui) folder from the git repo and a standalone [webapp](https://github.com/abel-research/ampscan_webapp) will soon be avilable. 
 
-Other packages perform the specific tasks of ampscan more comprehensively such as [mayavi](https://docs.enthought.com/mayavi/mayavi/index.html) for scientific data visualisation, [open3d](http://www.open3d.org/docs/getting_started.html) for alignment and registration techniques, and [gibbon](https://www.gibboncode.org/) for integration with finite element modelling. At this time, the package only contains limited mesh fixing tools that may be required for erroneous scans. Until such methods are implemented, the authors recommend [meshlab](http://www.meshlab.net/) which also contains methods for alignment and visualisation. An ampscan web-app is currently in development to further increase accessibility of these tools for clinicians and researchers working across the growing field of bespoke digital medical device design, and enable them to perform standardised, accessible and reliable analysis, and enhance patient experience.
+Other packages perform the specific tasks of ampscan more comprehensively such as mayavi for scientific data visualisation [@ramachandran2011mayavi], open3d for alignment and registration techniques [@Zhou2018], and GIBBON for integration with finite element modelling [@Moerman:GIBBON]. At this time, the package only contains limited mesh fixing tools that may be required for erroneous scans. Until such methods are implemented, the authors recommend meshlab which also contains methods for alignment and visualisation [@meshlab]. An ampscan web-app is currently in development to further increase accessibility of these tools for clinicians and researchers working across the growing field of bespoke digital medical device design, and enable them to perform standardised, accessible and reliable analysis, and enhance patient experience.
 
 # Acknowledgments 
 While the ampscan package was not funded directly, the authors would like to thank the following for their financial support:
