+++ joss/paper.bib
@@ -5,7 +5,8 @@
   volume={13},
   number={4},
   pages={551--557},
-  year={2005}
+  year={2005},
+  doi={10.1109/tnsre.2005.858459}
 }
 
 @article{Dickinson:2016,
@@ -15,7 +16,8 @@
   volume={53},
   number={2},
   pages={207--218},
-  year={2016}
+  year={2016},
+  doi={10.1682/jrrd.2014.10.0272}
 }
 
 @inproceedings{Dickinson:APOSM,
@@ -46,5 +48,46 @@
   title={Predictive prosthetic socket design. Part 1: population-based evaluation of transtibial prosthetic sockets by FEA-driven surrogate modelling},
   author={Steer, Joshua W and Browne, Martin and Worsley, Peter R and Dickinson, Alexander S},
   journal={Biomechanics and Modeling in Mechanobiology},
-  year={2019}
+  year={2019},
+  doi={10.1007/s10237-019-01195-5}
 }
+
+@article{Moerman:GIBBON,
+  title={GIBBON: The Geometry and Image-Based Bioengineering add-On},
+  author={Moerman, Kevin M},
+  journal={The Journal of Open Source Software},
+  year={2018},
+  doi={10.21105/joss.00506}
+}
+
+@article{Zhou2018,
+   author  = {Qian-Yi Zhou and Jaesik Park and Vladlen Koltun},
+   title   = {Open3D: A Modern Library for 3D Data Processing},
+   journal = {arXiv:1801.09847},
+   year    = {2018},
+}
+
+@article{ramachandran2011mayavi,
+  title={Mayavi: 3D Visualization of Scientific Data},
+  author={Ramachandran, P. and Varoquaux, G.},
+  journal={Computing in Science & Engineering},
+  volume={13},
+  number={2},
+  pages={40-51},
+  issn={1521-9615},
+  year={2011},
+  publisher={IEEE}
+  doi={10.1109/MCSE.2011.35}
+}
+
+@inproceedings{meshlab,
+  LocalChapterEvents:ItalChap:ItalianChapConf2008:129-136, 
+  booktitle = {Eurographics Italian Chapter Conference}, 
+  editor = {Vittorio Scarano and Rosario De Chiara and Ugo Erra}, 
+  title = {{MeshLab: an Open-Source Mesh Processing Tool}}, 
+  author = {Cignoni, Paolo and Callieri, Marco and Corsini, Massimiliano and Dellepiane, Matteo and Ganovelli, Fabio and Ranzuglia, Guido}, 
+  year = {2008}, 
+  publisher = {The Eurographics Association}, 
+  ISBN = {978-3-905673-68-5}, 
+  DOI = {10.2312/LocalChapterEvents/ItalChap/ItalianChapConf2008/129-136} }
+
