### 2020-02-09 9198feb
Fixed references in paper and added DOIs
