+++ joss/paper.bib
@@ -76,7 +76,7 @@
   pages={40-51},
   issn={1521-9615},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/MCSE.2011.35}
 }
 
