### 2020-02-09 1c34762
Fixed references in paper and added DOIs
