+++ joss/paper.bib
@@ -81,7 +81,6 @@
 }
 
 @inproceedings{meshlab,
-  LocalChapterEvents:ItalChap:ItalianChapConf2008:129-136, 
   booktitle = {Eurographics Italian Chapter Conference}, 
   editor = {Vittorio Scarano and Rosario De Chiara and Ugo Erra}, 
   title = {{MeshLab: an Open-Source Mesh Processing Tool}}, 
