### 2020-02-09 f4bac02
Fixed references in paper and added DOIs
