+++ joss/paper.bib
@@ -36,13 +36,18 @@
   place={Las Vegas, USA}
 }
 
-@inproceedings{Steer:ISPO,
+@article{Steer:ISPO,
   title={AmpScan: Open-source 3D scan shape analysis for prosthetics and orthotics},
   author={Steer, J. and Worsley, P. and Stocks, O. and Animashaun, O. and Dickinson, A.},
-  booktitle={Prosthetics and Orthotics International},
-  year={2019},
-  place={Kobe, Japan}
+  journal = {Prosthetics and Orthotics International},
+  volume = {43},
+  number = {1\_suppl},
+  pages = {94},
+  year = {2019},
+  doi:{10.1177/0309364619883197}
 }
+
+
 
 @article{Steer:BMMB,
   title={Predictive prosthetic socket design. Part 1: population-based evaluation of transtibial prosthetic sockets by FEA-driven surrogate modelling},
