### 2020-02-09 07bfa50
Fixed references in paper and added DOIs
