### 2020-02-09 c204d79
Fixed references in paper and added DOIs
