+++ joss/paper.bib
@@ -44,7 +44,7 @@
   number = {1\_suppl},
   pages = {94},
   year = {2019},
-  doi:{10.1177/0309364619883197}
+  doi = {10.1177/0309364619883197}
 }
 
 
