### 2020-02-09 28e0136
Fixed references in paper and added DOIs
