+++ joss/paper.bib
@@ -67,9 +67,8 @@
 
 @article{Zhou2018,
    author  = {Qian-Yi Zhou and Jaesik Park and Vladlen Koltun},
-   title   = {Open3D: A Modern Library for 3D Data Processing},
-   journal = {arXiv preprint},
-   doi = {arXiv:1801.09847},
+   title   = {Open3D: A Modern Library for 3D Data Processing},   
+   journal = {arXiv:1801.09847},
    year    = {2018},
 }
 
