### 2020-02-09 96bfa84
Increased contribution guidelines and code of conduct
