### 2020-02-09 2fa9b97
Added contributing guidelines to docs
