### 2020-02-09 9e707eb
Updated copyright year and contributing guidelines on the docs
