+++ setup.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Copyright: Joshua Steer 2019, Joshua.Steer@soton.ac.uk
+Copyright: Joshua Steer 2020, Joshua.Steer@soton.ac.uk
 """
 
 from setuptools import setup, find_packages
