+++ docs/LICENSE.rst
@@ -4,7 +4,7 @@ License
 MIT License
 -----------
 
-Copyright (c) 2019 Joshua Steer
+Copyright (c) 2020 Joshua Steer
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
