+++ docs/conf.py
@@ -35,7 +35,7 @@ on_rtd = os.environ.get('READTHEDOCS') == 'True'
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-# needs_sphinx = '1.0'
+needs_sphinx = '1.6'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
