+++ docs/requirements.txt
@@ -1,9 +1,9 @@
-sphinx==1.7.5
-sphinx_rtd_theme==0.4.3
+sphinx
+sphinx_rtd_theme
 numpy==1.16.4
 matplotlib==2.2.2
 scipy==1.2.1
-sphinxcontrib-napoleon==0.7
+sphinxcontrib-napoleon
 vtk==8.1.0
 PyQt5==5.13.0
 PyPDF2==1.26.0
