+++ docs/requirements.txt
@@ -1,13 +1,14 @@
+numpy>=1.12
+matplotlib>=2
+scipy>=1.2.1
+sphinxcontrib-napoleon
+vtk>=8
+PyQt5>=5,<=5.13
+PyPDF2>=1.20
+reportlab>=3
+
 sphinx
 sphinx_rtd_theme
-numpy==1.16.4
-matplotlib==2.2.2
-scipy==1.2.1
-sphinxcontrib-napoleon
-vtk==8.1.0
-PyQt5==5.13.0
-PyPDF2==1.26.0
-reportlab==3.5.23
 nbsphinx
 jupyter_sphinx
 git+https://github.com/abel-research/ampscan
