### 2020-02-09 #2060 @Kevin-Mattheus-Moerman
@JoshuaSteer
> Quick Q on testing, we've got our whole package linked up to travis CI (https://travis-ci.org/abel-research/ampscan), so I guess you are referring to enabling for someone to run these tests locally on install as well?

Apologies. Missed that. That is fine. I ticked the automated tests box.
