### 2020-02-09 e559bda
Updated docs index and repo readme so clear where testing located
