### 2020-02-09 #2060 @JoshuaSteer
@Kevin-Mattheus-Moerman

> @JoshuaSteer
>
> > Quick Q on testing, we've got our whole package linked up to travis CI (https://travis-ci.org/abel-research/ampscan), so I guess you are referring to enabling for someone to run these tests locally on install as well?
>
> Apologies. Missed that. That is fine. I ticked the automated tests box.

No worries, it wasn't immediately obvious on the docs/repo that we had an automated testing package, so I've edited that now.

Contribution guidelines, code of conduct, typos, acknowledged contributors, DOIs have all been changed as requested.

Great recommendations! We had planned to add in a HC smoothing algorithm (as you say, normal issues of volume loss with standard laplacian) and a 'true' volume calculation, so now is as good a time as any to include them.
