### 2020-02-09 3120f79
Added hc smoothing algorithm for improved volume preservation
