### 2020-02-09 51e1b56
Updated hc smoothing instructions and tests
