+++ tests/test_smoothing.py
@@ -19,6 +19,7 @@ class TestSmoothing(unittest.TestCase):
         from ampscan.core import AmpObject
         stl_path = get_path("stl_file_5.stl")
         self.amp = AmpObject(stl_path)
+        self.amp2 = AmpObject(stl_path)
 
     def test_smoothing_nans(self):
         """Tests that NaNs are properly dealt with by smooth method"""
@@ -31,8 +32,15 @@ class TestSmoothing(unittest.TestCase):
         # TODO check this is actually working properly
         poly1 = analyse.create_slices(self.amp, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
         print(analyse.est_volume(poly1))
+        
         self.amp.lp_smooth(1)
         poly2 = analyse.create_slices(self.amp, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
         print(analyse.est_volume(poly2))
         self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly2), delta=TestSmoothing.DELTA)
+        
+        self.amp2.hc_smooth(1)
+        poly3 = analyse.create_slices(self.amp, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
+        print(analyse.est_volume(poly3))
+        self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly3), delta=TestSmoothing.DELTA)
+        self.assertLess(analyse.est_volume(poly1)-analyse.est_volume(poly3), analyse.est_volume(poly1)-analyse.est_volume(poly2))
 
