### 2020-02-09 0e9506f
Updated hc smoothing tests
