### 2020-02-09 88c3220
Updated hc smoothing tests
