### 2020-02-09 e003a20
Updated hc smoothing tests
