+++ tests/test_smoothing.py
@@ -30,16 +30,16 @@ class TestSmoothing(unittest.TestCase):
     def test_smoothing_volume(self):
         """Tests that smoothing affects the volume within given acceptable range"""
         # TODO check this is actually working properly
-        poly1 = analyse.create_slices(self.amp, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
+        poly1 = analyse.create_slices(self.amp, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
         print(analyse.est_volume(poly1))
         
         self.amp.lp_smooth(20)
-        poly2 = analyse.create_slices(self.amp, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
+        poly2 = analyse.create_slices(self.amp, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
         print(analyse.est_volume(poly2))
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly2), delta=TestSmoothing.DELTA)
         
         self.amp2.hc_smooth(20)
-        poly3 = analyse.create_slices(self.amp2, [0.001, 0.999], 0.001, typ='norm_intervals', axis=2)
+        poly3 = analyse.create_slices(self.amp2, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
         print(analyse.est_volume(poly3))
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly3), delta=TestSmoothing.DELTA)
         self.assertLess(analyse.est_volume(poly1)-analyse.est_volume(poly3), analyse.est_volume(poly1)-analyse.est_volume(poly2))
