### 2020-02-09 95298e5
Updated hc smoothing tests
