+++ tests/test_smoothing.py
@@ -20,6 +20,7 @@ class TestSmoothing(unittest.TestCase):
         stl_path = get_path("stl_file_2.stl")
         self.amp = AmpObject(stl_path)
         self.amp2 = AmpObject(stl_path)
+        self.amp3 = AmpObject(stl_path)
 
     def test_smoothing_nans(self):
         """Tests that NaNs are properly dealt with by smooth method"""
@@ -31,16 +32,19 @@ class TestSmoothing(unittest.TestCase):
         """Tests that smoothing affects the volume within given acceptable range"""
         # TODO check this is actually working properly
         poly1 = analyse.create_slices(self.amp, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
-        print(analyse.est_volume(poly1))
+        vol1 = analyse.est_volume(poly1) 
+        print(vol1)
         
-        self.amp.lp_smooth(20)
-        poly2 = analyse.create_slices(self.amp, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
-        print(analyse.est_volume(poly2))
+        self.amp2.lp_smooth(20)
+        poly2 = analyse.create_slices(self.amp2, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
+        vol2 = analyse.est_volume(poly2) 
+        print(vol2)
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly2), delta=TestSmoothing.DELTA)
         
-        self.amp2.hc_smooth(20)
-        poly3 = analyse.create_slices(self.amp2, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
-        print(analyse.est_volume(poly3))
+        self.amp3.hc_smooth(20)
+        poly3 = analyse.create_slices(self.amp3, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
+        vol3 = analyse.est_volume(poly3) 
+        print(vol3)
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly3), delta=TestSmoothing.DELTA)
-        self.assertLess(analyse.est_volume(poly1)-analyse.est_volume(poly3), analyse.est_volume(poly1)-analyse.est_volume(poly2))
+        self.assertLess(vol1-vol3, vol1-vol2)
 
