### 2020-02-09 9b8287e
Updated hc smoothing tests
