### 2020-02-09 18024c6
Updated hc smoothing tests
