+++ tests/test_smoothing.py
@@ -31,18 +31,18 @@ class TestSmoothing(unittest.TestCase):
     def test_smoothing_volume(self):
         """Tests that smoothing affects the volume within given acceptable range"""
         # TODO check this is actually working properly
-        poly1 = analyse.create_slices(self.amp, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
+        poly1 = analyse.create_slices(self.amp, [0.02, 0.98, 0.005, typ='norm_intervals', axis=2)
         vol1 = analyse.est_volume(poly1) 
         print(vol1)
         
         self.amp2.lp_smooth(20)
-        poly2 = analyse.create_slices(self.amp2, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
+        poly2 = analyse.create_slices(self.amp2, [0.02, 0.98], 0.005, typ='norm_intervals', axis=2)
         vol2 = analyse.est_volume(poly2) 
         print(vol2)
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly2), delta=TestSmoothing.DELTA)
         
         self.amp3.hc_smooth(20)
-        poly3 = analyse.create_slices(self.amp3, [0.01, 0.99], 0.005, typ='norm_intervals', axis=2)
+        poly3 = analyse.create_slices(self.amp3, [0.02, 0.98], 0.005, typ='norm_intervals', axis=2)
         vol3 = analyse.est_volume(poly3) 
         print(vol3)
         # self.assertAlmostEqual(analyse.est_volume(poly1), analyse.est_volume(poly3), delta=TestSmoothing.DELTA)
