### 2020-02-09 9b151c2
Updated hc smoothing tests
