### 2020-02-09 9212a79
Updated hc smoothing tests
