### 2020-02-10 #2060 @Kevin-Mattheus-Moerman
@JoshuaSteer thanks for working on my suggestions. In relation to testing perhaps instruct users to set up Jupyter and run the notebooks. Is this how you intend users to run the tutorials? Is this written somewhere already?

Some minor bits:
- [ ] Check typos in the README, see these: Licesnse, founde, and here here
- [ ] I suggest linking to your contributing guidelines in the README
- [ ] A super minor nitpicky point. Perhaps capitalize `laplacian` throughout documentation (I see laplacian  in the fundamentals demo).
- [ ] Did you see the authorship/contributor acknowledgement point I made above ☝️
