### 2020-02-11 #2060 @JoshuaSteer
@Kevin-Mattheus-Moerman Good idea about the Jupyter notebooks for testing. I'll add in a section which describes to process to the user.

I've added in the HC smoothing algorithm, with a test on Travis and it's included in the fundamentals notebook where you can compare the shrinkage relative to the Laplacian.

Omar has been acknowledged within the paper now. Felt this was more appropriate than as an author considering relative contribution.

Thanks for picking up on those details. I'll let you know when they're done!
