### 2020-02-13 #2060 @danasolav
@JoshuaSteer, have you tried a fresh installation of the package on Windows 10? I have been struggling in the last two days to get it running with no success. I keep having issues with dependencies. For example, it seems like installing ampscan already includes installation of vtk, but in the readme you ask to pre-install vtk, so I got an error when it unsuccessfully tried to uninstall the existing vtk. After fixing that I still had a lot of other issues. Any suggestions besides installing Linux?
