### 2020-02-16 #2060 @JoshuaSteer
@danasolav I work using Windows 10, and my build works fine at the moment. However, the recent versions of PyQt and VTK appear to be throwing some installation issues, so it may be best to adjust the installation instructions to just use pip rather than conda. I'll run some fresh installs on my machine to see if any issues. Can you send me some screenshots?
