### 2020-02-19 #2060 @JoshuaSteer
@danasolav I've run some fresh installs on Windows 10 without problems, if you have a traceback of the error that'd be great
