### 2020-02-19 #2060 @danasolav
@JoshuaSteer Thanks. I'll try again today and send you some screenshots and traceback. as far as I remember the errors where related to vtk, PyQt5, and PyPDF2.
Can you share which environment you're using and if used conda or pip to install the packages this time?
