### 2020-02-23 #2060 @danasolav
@JoshuaSteer, here are my comments/questions (mostly referring to the GUI):

 [ ] Could you add some general information in the readme about what the toolbox is useful for. For example- why would aligning to shapes could be used for? What are the analysis capabilities?

 [ ] Could you add instructions and explanations for using the GUI? For example, in Align: What is Static and Moving, how do you use the different options (Centre, ICP, translation/rotation) and in which order? I noticed for example that trying to use Centre or translations/rotations after ICP doesn’t work.

 [ ] In my experience, ICP can be very sensitive to the initial position and orientation of the two shapes. Is that what the rotations and translations are for? to bring the shapes close together before ICP?

 [ ] Could you add a coordinate system to the visualization in Align? Without it it’s much harder to use the translations/rotations and there is also no scale for the size of the limb.

 [ ] Could you add a colorbar to the Registration results and an explanation in the readme as to what is the visualized measure?

 [ ] Could you add an example file that can be loaded in FE Analysis? Without it, it’s not clear what Analyse in the GUI does. I’m also confused because the documentation says that the analyse module does the slicing and volume calculation but in the GUI “Analyse” requires an FEA object.

 [ ] I couldn't find an explanation anywhere as to what Kinetic Measurement is?

 [ ] I tried using Generate Measurements after picking the mid patella point but it caused the GUI to crash and close. (FileNotFoundError: [Errno 2] No such file or directory: 'res\\Measurements Template.pdf'). I see this file in the res folder so I’m not sure why it fails to obtain it. Anyway, it looks like the template file doesn't contain volume. So is there a way to obtain volume measurement using the GUI?

 [ ] In Registration, why are the vertices morphed onto the other shape? Is it for having a direct vertex-to-vertex displacement value? There are also ways to compute it when the vertices do not correspond.

 [ ] I ran some tests using different STLs which are all originated from the same scan but in different rotations/ different mesh density/ top edge trimmed differently. I found that sometimes the ICP algorithm fails if I run it without making the two shapes very close first. These are the STLs I used for testing, in case you want to try them: https://www.dropbox.com/sh/6ubuj01y2a6cp6a/AADfI-ODmkXbR5Qf10X7v1TEa?dl=0
