### 2020-02-26 #2060 @JoshuaSteer
@danasolav Thanks for all the suggestions. A few things we've got in the pipeline to fix. There's a couple of legacy elements of the GUI which need removing as well which you picked up on.

I'll start working through them and let you know when they're updated

Big thanks for providing your STL files for testing - very helpful!
