### 2020-03-09 d79d76a
Updated readme to include more detailed overview of ampscan
