### 2020-03-09 d089c36
Removed legacy fields of kinetic measurements and FE from the GUI
