### 2020-03-09 513d827
Added triad to GUI
