### 2020-03-09 e58baf1
Capitalised Laplacian in docs and comments
