### 2020-03-09 8063a38
Capitalised Laplacian in docs and comments
