### 2020-03-09 202a995
Updated creation of measurements file!
