### 2020-03-09 61a364d
Updated registration show scalar bar visualised in the GUI
