### 2020-03-09 #2060 @JoshuaSteer
Thanks @Kevin-Mattheus-Moerman

@danasolav exposed some good bugs to tackle in the GUI. I've done a fairly sizeable restructure based on this. Should have edits finished in next day or so
