+++ gui/ampscanGUI.py
@@ -309,10 +309,9 @@ class ampscanGUI(QMainWindow):
             if hasattr(self, 'regCont'):
                 self.regCont.getNames()
             #im = []
-            if self.regCont.tick.isChecked() is True:
-                reg.actor.setScalarRange([-10,10])
-                reg.actor.setShading(False)
-                analyse.CMapOut(reg, colors=self.CMapN2P)
+            reg.actor.setScalarRange([-5,5])
+            reg.actor.setShading(False)
+            analyse.CMapOut(reg, colors=self.CMapN2P)
                 # reg.plotResults(name="distributionofshapevariance.png")
             self.display()  # Reset which objects are displayed
             self.renWin.setScalarBar(reg.actor, title = 'Shape difference, mm')
