### 2020-03-10 98dc7d6
Updated gui instructions
