### 2020-03-11 #2060 @JoshuaSteer
Thanks for all your great comments. It's really helped improve and focus the package:

@Kevin-Mattheus-Moerman all your following have now been sorted
- Check typos in the README, see these: Licesnse, founde, and here here
- I suggest linking to your contributing guidelines in the README
- A super minor nitpicky point. Perhaps capitalize laplacian throughout documentation (I see laplacian in the fundamentals demo).

@danasolav

-  Could you add some general information in the readme about what the toolbox is useful for. For example- why would aligning to shapes could be used for? What are the analysis capabilities?
_I've added a whole section to the readme which gives a high level overview of the toolbox and the capabilities of the individual modules_
-  Could you add instructions and explanations for using the GUI? For example, in Align: What is Static and Moving, how do you use the different options (Centre, ICP, translation/rotation) and in which order? I noticed for example that trying to use Centre or translations/rotations after ICP doesn’t work.
_I've added a new guide on the [readthedocs page](https://ampscan.readthedocs.io/en/latest/examples/gui_instructions.html)_
-  In my experience, ICP can be very sensitive to the initial position and orientation of the two shapes. Is that what the rotations and translations are for? to bring the shapes close together before ICP?
_Yes, and in our experience as we are performing ICP on dissimilar shapes, you may want to refine the automated alignment based upon your protocol_
-  Could you add a coordinate system to the visualization in Align? Without it it’s much harder to use the translations/rotations and there is also no scale for the size of the limb.
_Great suggestion,Triad now added to the GUI_
-  Could you add a colorbar to the Registration results and an explanation in the readme as to what is the visualized measure?
_Colourbar now appears in the GUI_
-  Could you add an example file that can be loaded in FE Analysis? Without it, it’s not clear what Analyse in the GUI does. I’m also confused because the documentation says that the analyse module does the slicing and volume calculation but in the GUI “Analyse” requires an FEA object.
-  I couldn't find an explanation anywhere as to what Kinetic Measurement is?
_Both of these should not have been here, they are legacy features from before we streamlined the software. Thanks for pointing them out as they've now been removed_
-  I tried using Generate Measurements after picking the mid patella point but it caused the GUI to crash and close. (FileNotFoundError: [Errno 2] No such file or directory: 'res\Measurements Template.pdf'). I see this file in the res folder so I’m not sure why it fails to obtain it. Anyway, it looks like the template file doesn't contain volume. So is there a way to obtain volume measurement using the GUI?
_It fails if the GUI wasn't booted from the correct folder. It should now always work as I've improved the file path manger for the res file_
-  In Registration, why are the vertices morphed onto the other shape? Is it for having a direct vertex-to-vertex displacement value? There are also ways to compute it when the vertices do not correspond.
_Which methods are you referring to specifically, for our purposes we find registration works well - often we compare coarse and fine meshes which require a point-to-plane style mesh_
-  I ran some tests using different STLs which are all originated from the same scan but in different rotations/ different mesh density/ top edge trimmed differently. I found that sometimes the ICP algorithm fails if I run it without making the two shapes very close first. These are the STLs I used for testing, in case you want to try them: https://www.dropbox.com/sh/6ubuj01y2a6cp6a/AADfI-ODmkXbR5Qf10X7v1TEa?dl=0
_Thanks for these, I didn't have these problems when I ran the algorithms on the various combinations, do you have an output of the error message?_
