### 2020-03-29 #2060 @danasolav
Sorry for the delay. @JoshuaSteer- thanks for answering my comments. I've updated my checklist, and the paper is ready for publication as far as I'm concerned.
