### 2020-03-31 abel-research/ampscan#53 @Kevin-Mattheus-Moerman
Exact volume computation without need for slicing
Volume computation does not require slicing. And by using slicing it becomes less accurate than it can be (as shown in the coarse/fine comparison). Have you thought about exact volume computation based on Gauss divergence theorem (here is a MATLAB implementation https://github.com/gibbonCode/GIBBON/blob/master/lib/triSurfVolume.m, note for quads you can first convert to triangular faces by slashing the quads). You probably need to close over the top of the scan though (although that should not be too hard).
