### 2020-03-31 #2060 @Kevin-Mattheus-Moerman
@JoshuaSteer thanks for making those changes. I think you addressed all my main points. I was wondering if you considered my suggesting of using the Gaussian divergence theorem for volume computation (which provides the exact volume and does not rely on slicing). This is not a requirement, you can argue you are happy with the slicing based method for now, I just wanted to know if you worked on this.
I've posted a related issue here: https://github.com/abel-research/ampscan/issues/53 in case you could consider this as a future development.
