### 2020-03-31 a049d08
Added in method to calculate the volume of a closed surface
