### 2020-03-31 8bf53af
Added in method to calculate the volume of a closed surface
