### 2020-03-31 f20dc83
Updated the volume calculation with hole filling
