+++ ampscan/analyse/__init__.py
@@ -1,4 +1,4 @@
-from .analyse import (create_slices, calc_perimeter, calc_widths, calc_csa, est_volume, 
+from .analyse import (calc_volume_closed, create_slices, calc_perimeter, calc_widths, calc_csa, est_volume, 
                       visualise_slices, plot_slices, MeasurementsOut, CMapOut)
 from .output import getPDF, generateRegBinsCsv, generateRegCsv
 
