### 2020-03-31 ace0039
Updated the volume calculation with hole filling
