+++ ampscan/analyse/analyse.py
@@ -20,21 +20,25 @@ import os
 # The file path used in doc examples
 filename = os.path.join(os.getcwd(), "tests", "stl_file.stl")
 
-def calc_volume_closed(amp_in, return_closed=True):
+def calc_volume_closed(amp_in, return_closed=False):
     r"""
-    Calculates the volume of a closed surface. If the surface is not closed the algorithm estimates the filled in holes 
+    Calculates the volume of a closed surface. If the surface is not closed the algorithm fills in the holes
+
+
 
     Parameters
     ----------
     amp: AmpObject 
         The AmpObject to analyse
-    fill: bool, default False
-        Indicate whether to fill the hole of an open surface. 
+    return_closed: bool, default False
+        Indicate whether to return the shape with holes filled in  
 
     Returns
     -------
     vol: float
         The volume of the AmpObject
+    amp: AmpObject
+        If 
     """
     amp = AmpObject({
         'vert': amp_in.vert.copy(),
@@ -42,18 +46,23 @@ def calc_volume_closed(amp_in, return_closed=True):
         'values': amp_in.values.copy(),
     })
     amp.calcStruct()
+    # Fill in the holes
     while (amp.faceEdges == -99999).sum() != 0: 
+        # Find the edges which are only conected to one face
         edges = (amp.faceEdges == -99999).sum(axis=1).astype(bool)
         edges = amp.edges[edges, :]
+        # Return the vert indicies for the loop
         vInd = logEuPath(edges)
+        # Calculate the mmidpoint 
         midpoint = amp.vert[vInd, :].mean(axis=0)
+        # Add in the new vertex
         amp.vert = np.r_[amp.vert, midpoint[None, :]]
         f0 = amp.vert.shape[0] - 1
+        # Add in each face using adjacent vertices in loop
         for f1, f2 in zip(vInd, np.roll(vInd, 1)):
             amp.faces = np.r_[amp.faces, [[f1, f0, f2]]]
+        # Update structure and check if any more holes (algorithm keeps going until all holes filled)
         amp.calcStruct()
-        # Add a new point at the midpoint of the loop
-        # Create faces from the edges
     # Calculate the area of each face in the array using vector cross product
     v01 = amp.vert[amp.faces[:, 1], :] - amp.vert[amp.faces[:, 0], :]
     v02 = amp.vert[amp.faces[:, 2], :] - amp.vert[amp.faces[:, 0], :]
@@ -61,7 +70,7 @@ def calc_volume_closed(amp_in, return_closed=True):
     area = 0.5 * np.sqrt(cp.sum(axis=1))
     # Get surface volume contributions 
     sVC = area * amp.vert[amp.faces, 2].mean(axis=1) * amp.norm[:, 2]
-    if return_closed == True:
+    if return_closed is True:
         return sVC.sum(), amp
     else:
         return sVC.sum()
