### 2020-03-31 994da11
Updated docs with hole filling algorithm
