### 2020-03-31 #2060 @JoshuaSteer
> @JoshuaSteer thanks for making those changes. I think you addressed all my main points. I was wondering if you considered my suggesting of using the Gaussian divergence theorem for volume computation (which provides the exact volume and does not rely on slicing). This is not a requirement, you can argue you are happy with the slicing based method for now, I just wanted to know if you worked on this.
> I've posted a related issue here: [abel-research/ampscan#53](https://github.com/abel-research/ampscan/issues/53) in case you could consider this as a future development.

This has now been done, thanks for the link to your algorithm as a reference, it worked great. I've updated the docs online, where you can see the result of the simple hole filling algorithm I wrote and how it changes the volume computation.

https://ampscan.readthedocs.io/en/latest/examples/analysis.html#Volume,-Cross-section-area-and-Widths

Hopefully that's now everything.
