### 2020-04-06 #2060 @Kevin-Mattheus-Moerman
@danasolav thanks again for reviewing this work. You've used tickboxes in this comment https://github.com/openjournals/joss-reviews/issues/2060#issuecomment-590065452 here. Would you be able to tick those boxes as well for completeness? Thanks
