### 2020-04-06 #2060 @JoshuaSteer
> @JoshuaSteer those changes look good. Thanks for implementing that volume computation. I have not more issues remaining. I'm happy (in my role as reviewer) to recommend this for acceptance in JOSS.

@Kevin-Mattheus-Moerman Excellent news! Thank you for all of yours and @danasolav work on this!
