+++ joss/paper.md
@@ -30,19 +30,19 @@ bibliography: paper.bib
 ---
 # Summary
 
-The increasing accessibility of surface scanners is giving users the ability to accurately digitise the 3D surface geometry of real world objects which may then be manufactured by CNC machining or 3D printing. In addition to hobbyist applications, these devices are being increasingly used within the medical field, for example prosthetics and orthotics clinics to capture the patients' individual limb geometry. These scans are imported into computer-aided design packages to generate patient-specific medical devices, such as prosthetic sockets, spinal braces or ankle-foot orthoses. This increasing digitsiation of patient data provides great potential for analysis in order to inform and improve clinical practice. While this has been an area of academic interest for several decades, clinical use is limited to date. One of the reasons for this is the lack of tools available for clinicians to analyse the geometry of their patient datasets. 
+The increasing accessibility of surface scanners is giving users the ability to accurately digitise the 3D surface geometry of real world objects which may then be manufactured by CNC machining or 3D printing. In addition to hobbyist applications, these devices are being increasingly used within the medical field, for example in prosthetics and orthotics clinics to capture the patients' individual limb geometry. These scans are imported into computer-aided design packages to generate patient-specific medical devices, such as prosthetic sockets, spinal braces or ankle-foot orthoses. This increasing digitsiation of patient data provides great potential for analysis in order to inform and improve clinical practice. While this has been an area of academic interest for several decades, clinical use is limited to date. One of the reasons for this is the lack of tools available for clinicians to analyse the geometry of their patient datasets. 
 
 Building on foundational work published by Zachariah, Sorenson and Sanders [@Sanders:2005], in a previously published paper [@Dickinson:2016] we detailed a method for comparing pairs of surface scans using alignment, registration and visualisation. This enabled extraction of key measurements for evaluation of the accuracy and reliability of clinically-used scanners. Further studies have also demonstrated using this package for comparing the consistency of casting techniques [@Dickinson:APOSM], statistical shape modeling across the population [@Steer:BMMB] and quantifying rectifications between the residual limb and the prosthetic socket [@Steer:AOPA]. 
 
-This method was written in MATLAB, although this was considered restrictive for other researchers to access as they may not possess a relevant license.  In order to maximise access to the developed techniques and improve performance, especially for 3D visualisation, the methods were rewritten from scratch within Python, including updated algorithms for each stage of the aforementioned process. 
+This method was written in MATLAB, although this was considered restrictive for other researchers to access as they may not possess a relevant license.  In order to maximise access to the developed techniques and improve performance, especially for 3D visualisation, the methods were rewritten from scratch within Python, including updated algorithms for each stage of the process. 
 
-The ampscan package has been designed with a range of user groups in mind, from practicing clinicians to biomedical engineering researchers, with an appreciation that they may not have an extensive background in coding. To this aim, the software has been developed in Python and leverages the commonly used libraries of NumPy, SciPy, matplotlib, vtk and pyqt. As such, full functionality of the software can be accessed without requiring additional installs. The core analysis of the package can be carried out both within a script-based environment, or alternatively with the supplied lightweight graphical user interface [@Steer:ISPO].  
[…]
