+++ joss/paper.bib
@@ -20,12 +20,12 @@
   doi={10.1682/jrrd.2014.10.0272}
 }
 
-@inproceedings{Dickinson:APOSM,
-  title={Assessing and benchmarking low cost 3D scanners for transtibial prosthetic socket design},
-  author={Dickinson, A S and Donovan-Hall, M K and Srors, S and Bou, K and Tech, A and Steer, J W and Metcalf, C D and Kheng, S and Worsley, P R},
-  booktitle={Asian Prosthetics and Orthotics Scientific Meeting},
-  year={2018},
-  place={Bangkok, Thailand}
+@article{Dickinson:APOSM,
+  title={Selecting Appropriate 3D Scanning Technologies for Prosthetic Socket Design and Transtibial Residual Limb Shape Characterisation},
+  author={Dickinson A, Donovan-Hall M, Kheng S, Bou K, Tech A, Steer J, Metcalf C, Worsley P},
+  journal={engrXiv},
+  year={2020},
+  doi={0.31224/osf.io/s4kbn}
 }
 
 @inproceedings{Steer:AOPA,
