+++ joss/paper.bib
@@ -25,7 +25,7 @@
   author={Dickinson A, Donovan-Hall M, Kheng S, Bou K, Tech A, Steer J, Metcalf C, Worsley P},
   journal={engrXiv},
   year={2020},
-  doi={0.31224/osf.io/s4kbn}
+  doi={10.31224/osf.io/s4kbn}
 }
 
 @inproceedings{Steer:AOPA,
