### 2020-04-06 #2060 @JoshuaSteer
> @JoshuaSteer your work is about to be processed for acceptance in JOSS.
>
> * [ ]  Please thoroughly review your paper including the author names and affiliations one last time.
> * [ ]  Once you've read and updated your paper please post an archived version of your software on Zenodo. The Zenodo archived version's meta data, such as the title and author list, should match those of your paper! (these instructions might be helpful: https://guides.github.com/activities/citable-code/)
> * [ ]  Once you've archived a version there please report back the DOI of the archived version here.
> * [ ]  Can you confirm what the version tag is for the archived version of your software.

@Kevin-Mattheus-Moerman

> @JoshuaSteer thanks for working on my suggestions. In relation to testing perhaps instruct users to set up Jupyter and run the notebooks. Is this how you intend users to run the tutorials? Is this written somewhere already?
>
> Some minor bits:
>
> * [x]  Check typos in the README, see these: Licesnse, founde, and here here
> * [x]  I suggest linking to your contributing guidelines in the README
> * [x]  A super minor nitpicky point. Perhaps capitalize `laplacian` throughout documentation (I see laplacian  in the fundamentals demo).
> * [x]  Did you see the authorship/contributor acknowledgement point I made above ☝️

> @JoshuaSteer your work is about to be processed for acceptance in JOSS.
>
> * [x]  Please thoroughly review your paper including the author names and affiliations one last time.
> * [x]  Once you've read and updated your paper please post an archived version of your software on Zenodo. The Zenodo archived version's meta data, such as the title and author list, should match those of your paper! (these instructions might be helpful: https://guides.github.com/activities/citable-code/)
> * [x]  Once you've archived a version there please report back the DOI of the archived version here.
> * [x]  Can you confirm what the version tag is for the archived version of your software.

@Kevin-Mattheus-Moerman

DOI: 10.5281/zenodo.3741495
Version: v0.3.0
