### 2020-04-14 #2060 @JoshuaSteer
> @danasolav, @Kevin-Mattheus-Moerman many thanks for your reviews here and to @Kevin-Mattheus-Moerman for editing this submission too ✨
>
> @JoshuaSteer - your paper is now accepted into JOSS ⚡🚀💥

That's fantastic news. Thanks everyone for all your help reviewing this!
